build(core): bundle and auto release
