+++ global.ts
@@ -1,0 +1,7 @@
+/// <reference types="@repo/dts/common" />
+
+declare namespace NodeJS {
+  interface ProcessEnv {
+    LINT_STAGED_TYPE?: "format";
+  }
+}
