+++ packages/core/package.json
@@ -16,17 +16,6 @@
   },
   "license": "Apache-2.0",
   "type": "module",
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "default": "./dist/index.js"
-    },
-    "./**/*": {
-      "types": "./dist/**/*.d.ts",
-      "default": "./dist/**/*.js"
-    },
-    "./package.json": "./package.json"
-  },
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
   "files": [
