+++ packages/examples/tsconfig.app.json
@@ -1,10 +1,7 @@
 {
   "compilerOptions": {
     "paths": {
-      "@crescendolab/zodios-react-query-v5": [
-        "../../packages/core/src/index.ts"
-      ],
-      "@crescendolab/zodios-react-query-v5/*": ["../../packages/core/src/*"],
+      "@crescendolab/zodios-react-query-v5": ["../core/src/index.ts"],
       "@utils/*": ["./src/utils/*"]
     },
     "tsBuildInfoFile": "./node_modules/.tmp/tsconfig.app.tsbuildinfo"
